module-3-agent-architectures/migration-workflow-system/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
openai==0.28.1
python-dotenv==1.0.0
